refactor; move query to own module
