+++ cognite/pygen/__init__.py
@@ -4,8 +4,8 @@ generating SDKs.
 """
 
 from ._build import build_wheel
-from ._core.query_execution import QueryExecutor as _QueryExecutor
 from ._generator import generate_sdk, generate_sdk_notebook
+from ._query import QueryExecutor as _QueryExecutor
 from ._version import __version__
 from .utils.cdf import load_cognite_client_from_toml
