+++ cognite/pygen/_query/__init__.py
@@ -1,0 +1,3 @@
+from .query_execution import QueryExecutor
+
+__all__ = ["QueryExecutor"]
